// src/app/interface/clinic.interface.ts
export interface Clinic {
    name: string;
    addressLineOne: string;
    city: string;
    zipCode: string;
    stateOrRegion: string;
    country: string;
    phone: string;
    clinicEmail: string;
    website: string;
    language: string;
    contactPhone: string;
    first_name: string;
    last_name: string;
    contactName: string;
    email: string;
    contactEmail: string;
    position: string;
    status: string;
    created_at: string;
    created_by: string;
    addressLineTwo?: string;
    id?: number;
}